Build: Split config into two parts - blocks and packages
