+++ webpack.config.js
@@ -1,6 +1,7 @@
 /**
  * Internal dependencies
  */
+const blocksConfig = require( './tools/webpack/blocks' );
 const packagesConfig = require( './tools/webpack/packages' );
 
-module.exports = packagesConfig;
+module.exports = [ blocksConfig, packagesConfig ];
